frm_t_c

+++ ProyectoNido/WcfNido/clsEntidades/clsTarifario.cs
@@ -12,13 +12,13 @@ namespace clsEntidades
         public string Tipo { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
-        public byte Periodo { get; set; }
+        public int Periodo { get; set; }
         public decimal Valor { get; set; }
 
         public clsTarifario() { }
 
         public clsTarifario(int id, string tipo, string nombre, string descripcion,
-                            byte periodo, decimal valor)
+                            int periodo, decimal valor)
         {
             Id_Tarifario = id;
             Tipo = tipo;
